Try to megre from BOHDAN

+++ feature_extraction_parallel/corner_detection_parallel.h
@@ -13,7 +13,7 @@
 
 inline size_t NUMBER_OF_THREADS = 4;
 inline int BLOCK_SIZE = 100;
-constexpr double RESPONSE_THRESHOLD = 25000;
+constexpr float RESPONSE_THRESHOLD = 25000;
 
 struct bound2d {
     int start{}, end{};
@@ -38,12 +38,10 @@ using work_result = std::variant<bool, std::vector<cv::KeyPoint>, std::vector<st
 class CornerDetectionParallel {
 public:
     static cv::Mat custom_bgr2gray(cv::Mat& picture);
-    // static std::vector<cv::KeyPoint> non_maximum_suppression_worker(const std::vector<std::vector<double>> &R_values, const int& n_rows, const int& n_cols, const int& k, const int& N);
 
     static void direction_gradients_worker(const cv::Mat& picture, const interval& interval, cv::Mat& Jx, cv::Mat& Jy, cv::Mat& Jxy);
-    static void shitomasi_corner_detection_worker(const cv::Mat& Jx, const cv::Mat& Jy, const cv::Mat& Jxy, const interval& interval, const double& k, std::vector<std::vector<double>>& R_array);
-    // static std::vector<cv::KeyPoint> non_maximum_suppression_worker(std::vector<std::vector<double>> R_values, const int& n_rows, const int& n_cols, const int& k, const int& N);
-    static std::vector<cv::KeyPoint> non_maximum_suppression(const std::vector<std::vector<double>> &R_values, const int& n_rows, const int& n_cols, const int& k, const int& N);
+    static void shitomasi_corner_detection_worker(const cv::Mat& Jx, const cv::Mat& Jy, const cv::Mat& Jxy, const interval& interval, const float& k, std::vector<std::vector<float>>& R_array);
+    static std::vector<cv::KeyPoint> non_maximum_suppression(const std::vector<std::vector<float>> &R_values, const int& n_rows, const int& n_cols, const int& k, const int& N);
 
 
 };
